# aspnetcore/src/IO.Swagger/Controllers/TreatmentPlanApi.cs
/*
 * Healthcare
 *
 * This is a API built for demo purpose. About patients and their treatment plan.
 *
 * OpenAPI spec version: 1.0.3
 * Contact: m.naseem@outlook.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TreatmentPlanApiController : ControllerBase
    { 
        /// <summary>
        /// Get patient&#x27;s treatment plan
        /// </summary>
        /// <remarks>By passing in the patientId, you can get the  patient&#x27;s treamtment plan </remarks>
        /// <param name="authorization"></param>
        /// <param name="patientId">pass the patientId for looking up the database</param>
        /// <response code="200">search result matching criteria</response>
        /// <response code="400">bad input parameter</response>
        [HttpGet]
        [Route("/Tonic.Tims/Healthcare/1.0.0/patient/{patientId}/treatmentplan")]
        [ValidateModelState]
        [SwaggerOperation("GetTreatementPlan")]
        [SwaggerResponse(statusCode: 200, type: typeof(TreatmentPlan), description: "search result matching criteria")]
        public virtual IActionResult GetTreatementPlan([FromHeader][Required()]string authorization, [FromRoute][Required]string patientId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(TreatmentPlan));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);
            string exampleJson = null;
            exampleJson = "{\n  \"patientId\" : \"d290f1ee-6c54-4b01-90e6-d701748f0851\",\n  \"doctorId\" : \"d290f1ee-6c54-4b01-90e6-d701748f0851\",\n  \"prescription\" : [ \"None\", \"None\" ],\n  \"id\" : \"d290f1ee-6c54-4b01-90e6-d701748f0851\",\n  \"classification\" : [ \"Normal\", \"Normal\" ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<TreatmentPlan>(exampleJson)
                        : default(TreatmentPlan);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}