# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Healthcare",
    "description" : "This is a API built for demo purpose. About patients and their treatment plan.",
    "contact" : {
      "email" : "m.naseem@outlook.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.3"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/Tonic.Tims/Healthcare/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
...
  "paths" : {
...
    "/patient/{patientId}/treatmentplan" : {
      "get" : {
        "tags" : [ "Treatment plan" ],
        "summary" : "Get patient's treatment plan",
        "description" : "By passing in the patientId, you can get the  patient's treamtment plan\n",
        "operationId" : "getTreatementPlan",
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "patientId",
          "in" : "path",
          "description" : "pass the patientId for looking up the database",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search result matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TreatmentPlan"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "TreatmentPlan" : {
        "required" : [ "doctorId", "hospitalId", "id", "patientId", "patientclassification", "treatmentPlan" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "patientId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "doctorId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "classification" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "example" : "Normal",
              "enum" : [ "Underweight", "Normal", "Overweight", "Obese" ]
            }
          },
          "prescription" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "example" : "None",
              "enum" : [ "None", "Eat eggs", "Run 5 miles", "Run 15 miles" ]
            }
          }
        },
        "example" : {
          "patientId" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "doctorId" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "prescription" : [ "None", "None" ],
          "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "classification" : [ "Normal", "Normal" ]
        }
      }
    }
  }
}